paper: Directly use ChunkSystem.scheduleChunkLoad instead of ServerChunkCache#getChunkAtAsynchronously
The later converts the requested chunk status to ticket level and back, which results in STRUCTURE_STARTS getting requested instead of EMPTY (because ChunkHolder.getStatus(MCUtil.getTicketLevelFor(ChunkStatus.EMPTY)) returns STRUCTURE_STARTS). Which causes the chunks to get saved. This would lead to repeated full renders discovering more and more region files, but no more full chunks (rendered are would not increase). This is a bug in Paper, not squaremap, but by using the new chunk system directly instead of having getChunkAtAsynchronously do it for us, we can avoid the issue without any change to Paper.