web: block HTML in marker text
Wrap marker text in an HTMLElement text node prior to inserting it
into Leaflet. The browser treats this text as "verbatim" and does
not interpret any characters as HTML. The Leaflet
`DivOverlay.setContent()` method accepts the HTMLElement
directly [1] and will insert it into the DOM tree as-is [2].

This conforms to Mozilla's recommendations for safely inserting
external content [3] and should block all HTML in map markers.

Also add some CSS so that whitespace and linebreaks in marker
text are treated literally.

References

1. https://leafletjs.com/reference.html#divoverlay-setcontent

2. https://github.com/Leaflet/Leaflet/blob/f10b44b3afcd079febe5219f84dc67c68b379b5e/src/layer/DivOverlay.js#L280

3. https://developer.mozilla.org/en-US/docs/Mozilla/Add-ons/WebExtensions/Safely_inserting_external_content_into_a_page